refactor: folder restructure, README, debug file, typings, ...
