+++ src/Blockchain/index.ts
@@ -1,46 +1,47 @@
 import {
   MAINNET_NODES as BINANCE_MAINNET_NODES,
   TESTNET_NODES as BINANCE_TESTNET_NODES
-} from './nodes/binance';
-import { MAINNET_NODES as ETHEREUM_MAINNET_NODES } from './nodes/ethereum';
+} from './binance';
+import { MAINNET_NODES as ETHEREUM_MAINNET_NODES } from './ethereum';
 import {
   MAINNET_NODES as HARMONY_MAINNET_NODES,
   TESTNET_NODES as HARMONY_TESTNET_NODES
-} from './nodes/harmony';
+} from './harmony';
 import {
   MAINNET_NODES as FANTOM_MAINNET_NODES,
   TESTNET_NODES as FANTOM_TESTNET_NODES
-} from './nodes/fantom';
+} from './fantom';
 import {
   MAINNET_NODES as POLYGON_MAINNET_NODES,
   TESTNET_NODES as POLYGON_TESTNET_NODES
-} from './nodes/polygon';
+} from './polygon';
 import {
   MAINNET_NODES as IOTEX_MAINNET_NODES,
   TESTNET_NODES as IOTEX_TESTNET_NODES
-} from './nodes/iotex';
+} from './iotex';
 import {
   MAINNET_NODES as AVALANCHE_MAINNET_NODES,
   TESTNET_NODES as AVALANCHE_TESTNET_NODES
-} from './nodes/avalanche';
+} from './avalanche';
 import {
   MAINNET_NODES as BITTORRENT_MAINNET_NODES,
   TESTNET_NODES as BITTORRENT_TESTNET_NODES
-} from './nodes/bittorrent';
+} from './bittorrent';
 import {
   MAINNET_NODES as ONTOLOGY_MAINNET_NODES,
   TESTNET_NODES as ONTOLOGY_TESTNET_NODES
-} from './nodes/ontology';
+} from './ontology';
 import {
   EVENT_MAINNET_NODES as TRON_EVENT_MAINNET_NODES,
   FULLHOST_MAINNET_NODES as TRON_FULLHOST_MAINNET_NODES,
   TESTNET_NODES as TRON_TESTNET_NODES
-} from './nodes/tron';
+} from './tron';
 import {
   MAINNET_NODES as ICON_MAINNET_NODES,
   TESTNET_NODES as ICON_TESTNET_NODES
-} from './nodes/icon';
-export const NODES: any = {
+} from './icon';
+
+export const NODES: Record<string, Record<string, string[]>> = {
   avalanche: {
     mainnet: AVALANCHE_MAINNET_NODES,
     testnet: AVALANCHE_TESTNET_NODES
